refact: use StateFlow value delegates from :ext:coroutines
